perf(client): stop the basic properties tab from pulling in CodeMirror
The always-loaded basic properties tab imported CodeMimeTypesList from the
code note options page, which statically imports the full CodeMirror index
(including the vim keybindings extension) for its editor preview and theme
list. The MIME type list only needs the lightweight syntax highlighting
tables, so it now lives in its own module together with its styles.
