+++ apps/client/src/widgets/type_widgets/options/code_mime_types_list.css
@@ -1,0 +1,14 @@
+.options-mime-types {
+    list-style-type: none;
+    column-width: 250px;
+    padding-inline-start: 0;
+
+    & > section {
+        margin-bottom: 1em;
+        break-inside: avoid-column;
+    }
+
+    ul {
+        padding-inline-start: 0;
+    }
+}
